Move from BSD to MIT license for consistency
This moves the Expo client project as well as several of our other projects from the BSD license to the MIT license. The licenses are fairly comparable and are compatible. This (a) consistently licenses our latest code using the same license and (b) uses the same license as React and React Native. There should be no material impact to developers using Expo aside from one fewer license type they need to think about; generally standard BSD and MIT are well-accepted.

fbshipit-source-id: 020b49e